+++ backend/StudyRoom.API/Repositories/IStudySessionRepository.cs
@@ -7,8 +7,8 @@ public interface IStudySessionRepository
     Task AddAsync(StudySession session);
     Task UpdateAsync(StudySession session);
     Task<List<StudySession>> GetByUserIdAsync(Guid userId);
-    Task<int> GetTotalStudyMinutesAsync(Guid userId);
+    Task<decimal> GetTotalStudyMinutesAsync(Guid userId);
     Task<int> GetSessionsCompletedAsync(Guid userId);
-    Task<int> GetWeeklyStudyMinutesAsync(Guid userId);
+    Task<decimal> GetWeeklyStudyMinutesAsync(Guid userId);
     Task<int> GetCurrentStreakAsync(Guid userId);
 }
